refactor!: add stronger typing and linting, fix some minor bugs
Warning: this is a breaking change. Read on to understand the reasoning for
this, as well as more details.

refactor: add stronger typing
===
This change adds stronger typing to the project. This means avoiding the use of
the `any` type as much as possible, and using more specific types and generics
instead.  Where this is not possible, we use the `unknown` type, which is a
better alternative to "any" (`any` allows you to do anything with the value,
while `unknown` forces you to narrow the type before you can use it).

As a user of the library, I have wished for stronger typing for a while, and I
think it will help users of the library write better code.

BREAKING CHANGE: I am not sure what the exact implication of this is, but it's
possible that users have relied on the fact that the `any` type was used. Many
of these usages might now be type check errors.

refactor: add typescript-eslint to detect more anti-patterns
===
This change adds typescript-eslint to the project, which is a set of rules
that detect more anti-patterns than the default eslint rules. This can help us
write better code, and avoid some bugs.

refactor: NestedArray<T> type allows arbitrary nesting
===
The old type used to allow for 8 levels of nesting, while the new type allows
for arbitrary nesting. This should be a backward compatible change.

fix: some minor bugs
===
Mostly some minor bugs that were detected by the stricter linting.

test: add some tests
===
I added tests for `manyTill` and `manyBetween`, which were missing. I also
upgraded the jasmine related packages.
Some test related code is not typed as strongly as it could be, but this is
perhaps not worth the effort right now.

chore: regenerate yarn.lock