chore: add prettier-plugin-organize-imports
This is a plugin that makes prettier reformat the import sections of
source code files. Its goal is:

> Make sure that your import statements stay consistent no matter who
> writes them and what their preferences are.

Its readme says it uses the `organizeImports` feature of the typescript
language api, which is the same thing as the "Organize Imports" code
action in VS Code.

In a way, it extends prettier's philosophy of "let's agree on a coding
style that is enforced by tooling" to imports.

NOTE:
- if you run prettier when a file is saved, some unfinished imports will
  be destructively removed using this plugin - use the "undo" feature of
  your IDE to get them back in this case.
- if you also have an "automatically add imports" feature working, you
  can pretty much not worry about imports at all 99% of the time. In my
  experience this sort of developer experience is excellent.
- adding this plugin will affect each phase where prettier itself is
  used - namely:
  - during development, meaning in the IDE or on the command line
  - as a git pre-commit hook via `lint-staged`
  - every time prettier is used in CI

https://www.npmjs.com/package/prettier-plugin-organize-imports