fix: stringLen failing on repeated parsing
With my understanding, the old and deprecated "".substr method takes the
parameters (startIndex and length), returning that many characters. But
the new "".substring, while otherwise appearing similar, takes
(startIndex, endIndex) which provides very different results.

The difference can not be seen unless specific numbers are given - this
was not caught in the tests :)

Fixes #107